refactor: Member 서비스 분류 #25
