+++ src/main/java/com/example/everguide/config/SchedulerConfig.java
@@ -1,6 +1,6 @@
 package com.example.everguide.config;
 
-import com.example.everguide.service.member.MemberCommandService;
+import com.example.everguide.service.member.MemberService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -11,11 +11,11 @@ import org.springframework.stereotype.Component;
 @Slf4j
 public class SchedulerConfig {
 
-    private final MemberCommandService memberCommandService;
+    private final MemberService memberService;
 
     @Scheduled(cron = "0 0 0/1 * * ?")
     public void run() {
 
-        memberCommandService.updateRedis();
+        memberService.updateRedis();
     }
 }
