+++ src/main/java/com/example/everguide/service/member/MemberServiceImpl.java
@@ -33,21 +33,18 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import com.example.everguide.domain.enums.Role;
 
 import com.example.everguide.repository.MemberRepository;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
 
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
 import java.util.*;
 
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class MemberCommandServiceImpl implements MemberCommandService {
+public class MemberServiceImpl implements MemberService {
 
     @Value("${spring.security.oauth2.client.registration.kakao.client-id}")
     private String kakaoClientId;
@@ -60,298 +57,22 @@ public class MemberCommandServiceImpl implements MemberCommandService {
 
     private final MemberRepository memberRepository;
     private final BookmarkRepository bookmarkRepository;
+    private final MailService mailService;
+
     private final BCryptPasswordEncoder bCryptPasswordEncoder;
     private final RedisUtils redisUtils;
     private final RedisTemplate<String, Object> redisTemplate;
     private final JWTUtil jwtUtil;
-    private final MailService mailService;
     private final SecurityUtil securityUtil;
 
     @Override
-    @Transactional
-    public Boolean cookieToHeader(HttpServletRequest request, HttpServletResponse response) {
+    public Boolean checkOriginalPwd(MemberRequest.ChangePwdDTO changePwdDTO) {
 
-        String refresh = null;
-        Cookie[] cookies = request.getCookies();
-        for (Cookie cookie : cookies) {
-
-            if (cookie.getName().equals("refresh")) {
-
-                refresh = cookie.getValue();
-            }
-        }
-
-        if (refresh == null) {
-
-            throw new MemberBadRequestException(ErrorStatus._TOKEN_NULL.getMessage());
-        }
-
-        try {
-            jwtUtil.isExpired(refresh);
-
-        } catch (ExpiredJwtException e) {
-
-            throw new MemberBadRequestException(ErrorStatus._TOKEN_EXPIRED.getMessage());
-        }
-
-        String category = jwtUtil.getCategory(refresh);
-        if (!category.equals("refresh")) {
-
-            throw new MemberBadRequestException("Refresh Token이 아닙니다.");
-        }
-
-        String userId = jwtUtil.getUserId(refresh);
-        String role = jwtUtil.getRole(refresh);
-        String social = jwtUtil.getSocial(refresh);
-
-        // Redis에 저장되어 있는지 확인
-        String currentLocalRefresh = redisUtils.getLocalRefreshToken(userId);
-        if (!refresh.equals(currentLocalRefresh)) {
-
-            throw new MemberBadRequestException(ErrorStatus._LOCAL_INVALID_TOKEN.getMessage());
-        }
-
-        //make new JWT
-        String access = jwtUtil.createJwt(userId, role, social, "access", 60000*10L);
-
-        //response
-        response.setHeader("Authorization", "Bearer " + access);
-        response.setStatus(HttpStatus.OK.value());
-
-
-        return Role.getRole(role).equals(Role.ROLE_MEMBER);
-    }
-
-    @Override
-    @Transactional
-    public boolean reissue(HttpServletRequest request, HttpServletResponse response) {
-
-        String refresh = null;
-        Cookie[] cookies = request.getCookies();
-        for (Cookie cookie : cookies) {
-
-            if (cookie.getName().equals("refresh")) {
-
-                refresh = cookie.getValue();
-            }
-        }
-
-        if (refresh == null) {
-
-            throw new MemberBadRequestException(ErrorStatus._TOKEN_NULL.getMessage());
-        }
-
-        try {
-            jwtUtil.isExpired(refresh);
-
-        } catch (ExpiredJwtException e) {
-
-            throw new MemberBadRequestException(ErrorStatus._TOKEN_EXPIRED.getMessage());
-        }
-
-        String category = jwtUtil.getCategory(refresh);
-        if (!category.equals("refresh")) {
-
-            throw new MemberBadRequestException("Refresh Token이 아닙니다.");
-        }
-
-        String userId = jwtUtil.getUserId(refresh);
-        String role = jwtUtil.getRole(refresh);
-        String social = jwtUtil.getSocial(refresh);
-
-        // Redis에 저장되어 있는지 확인
-        String currentLocalRefresh = redisUtils.getLocalRefreshToken(userId);
-        if (!refresh.equals(currentLocalRefresh)) {
-
-            throw new MemberBadRequestException(ErrorStatus._LOCAL_INVALID_TOKEN.getMessage());
-        }
-
-        //make new JWT
-        String newAccess = jwtUtil.createJwt(userId, role, social, "access", 60000*10L);
-        String newRefresh = jwtUtil.createJwt(userId, role, social, "refresh", 60000*60*24L);
-
-        redisUtils.changeLocalRefreshToken(userId, newRefresh, 60000*60*24L);
-
-        //response
-        response.setHeader("Authorization", "Bearer " + newAccess);
-        response.addCookie(createCookie("refresh", newRefresh));
-        response.setStatus(HttpStatus.OK.value());
-
-        return true;
-    }
-
-    @Override
-    @Transactional
-    public boolean localSignUp(MemberRequest.SignupDTO signupDTO) {
-
-        String name = signupDTO.getName();
-        String birth = signupDTO.getBirth();
-        String phoneNumber = signupDTO.getPhoneNumber();
-        String email = signupDTO.getEmail();
-        String password = signupDTO.getPassword();
-        String userId = "LOCAL_" + email;
-
-        checkEmailExist(userId);
-
-        checkPhoneNumberExist(phoneNumber);
-
-//        checkSmsVerify(phoneNumber);
-
-        Member member = Member.builder()
-                .name(name)
-                .birth(LocalDate.parse(birth, DateTimeFormatter.BASIC_ISO_DATE))
-                .phoneNumber(phoneNumber)
-                .email(email)
-                .password(bCryptPasswordEncoder.encode(password))
-                .role(Role.ROLE_MEMBER)
-                .providerType(ProviderType.LOCAL)
-                .userId(userId)
-                .build();
-
-        memberRepository.save(member);
-
-        return true;
-    }
-
-    private Boolean checkSmsVerify(String phoneNumber) {
-
-        String savedCode = redisUtils.getSmsAuthCode(phoneNumber);
-        String verifyCode = redisUtils.getSmsAuthCodeVerify(phoneNumber);
-
-        if (savedCode == null || verifyCode == null) {
-            throw new MemberBadRequestException("인증코드를 찾을 수 없습니다.");
-        }
-
-        if (!savedCode.equals(verifyCode)) {
-            throw new MemberBadRequestException("인증코드가 일치하지 않습니다.");
-        }
-
-        return true;
-    }
-
-    private Boolean checkEmailExist(String userId) {
-
-        if (memberRepository.existsByUserId(userId)) {
-            throw new MemberBadRequestException("이미 존재하는 이메일입니다.");
-
-        } else {
-            return true;
-        }
-    }
-
-    private Boolean checkPhoneNumberExist(String phoneNumber) {
-
-        if (memberRepository.existsByPhoneNumber(phoneNumber)) {
-            throw new MemberBadRequestException("이미 존재하는 전화번호입니다.");
-
-        } else {
-            return true;
-        }
-    }
-
-    @Override
-    @Transactional
-    public MemberResponse.SignupAdditionalDTO getSignupAdditionalInfo(HttpServletRequest request, HttpServletResponse response) {
-
-        String authorization = request.getHeader("Authorization");
-
-        if (authorization == null || !authorization.startsWith("Bearer ")) {
-
-            throw new MemberBadRequestException(ErrorStatus._NO_TOKEN.getMessage());
-        }
-
-        String accessToken = authorization.split(" ")[1];
-
-        try {
-            jwtUtil.isExpired(accessToken);
-        } catch (ExpiredJwtException e) {
-
-            throw new MemberBadRequestException(ErrorStatus._TOKEN_EXPIRED.getMessage());
-        }
-
-        String category = jwtUtil.getCategory(accessToken);
-
-        if (!category.equals("access")) {
-
-            throw new MemberBadRequestException("Access Token이 아닙니다.");
-        }
-
-        String userId = jwtUtil.getUserId(accessToken);
+        String userId = securityUtil.getCurrentUserId();
 
         Member member = memberRepository.findByUserId(userId).orElseThrow(EntityNotFoundException::new);
 
-        return MemberResponse.SignupAdditionalDTO.builder()
-                .name(member.getName())
-                .birth(member.getBirth())
-                .phoneNumber(member.getPhoneNumber())
-                .email(member.getEmail())
-                .build();
-    }
-
-    @Override
-    @Transactional
-    public boolean registerSignupAdditionalInfo(
-            HttpServletRequest request, HttpServletResponse response,
-            MemberRequest.SignupAdditionalDTO signupAdditionalDTO) {
-
-        String authorization = request.getHeader("Authorization");
-
-        if (authorization == null || !authorization.startsWith("Bearer ")) {
-
-            throw new MemberBadRequestException(ErrorStatus._NO_TOKEN.getMessage());
-        }
-
-        String accessToken = authorization.split(" ")[1];
-
-        try {
-            jwtUtil.isExpired(accessToken);
-        } catch (ExpiredJwtException e) {
-
-            throw new MemberBadRequestException(ErrorStatus._TOKEN_EXPIRED.getMessage());
-        }
-
-        String category = jwtUtil.getCategory(accessToken);
-
-        if (!category.equals("access")) {
-
-            throw new MemberBadRequestException("Access Token이 아닙니다.");
-        }
-
-        String userId = jwtUtil.getUserId(accessToken);
-        String social = jwtUtil.getSocial(accessToken);
-
-        Member member = memberRepository.findByUserId(userId).orElseThrow(EntityNotFoundException::new);
-
-        member.setName(signupAdditionalDTO.getName());
-        LocalDate birth = LocalDate.parse(signupAdditionalDTO.getBirth(), DateTimeFormatter.BASIC_ISO_DATE);
-        member.setBirth(birth);
-        member.setPhoneNumber(signupAdditionalDTO.getPhoneNumber());
-        member.setEmail(signupAdditionalDTO.getEmail());
-        member.setRole(Role.ROLE_MEMBER);
-
-        memberRepository.save(member);
-
-        String access = jwtUtil.createJwt(userId, Role.ROLE_MEMBER.name(), social, "access", 60000*10L);
-        String refresh = jwtUtil.createJwt(userId, Role.ROLE_MEMBER.name(), social, "refresh", 60000*60*24L);
-
-        redisUtils.changeLocalRefreshToken(userId, refresh, 60000*60*24L);
-
-        response.addHeader("Authorization", "Bearer " + access);
-        response.addCookie(createCookie("refresh", refresh));
-        response.setStatus(HttpStatus.OK.value());
-
-        return true;
-    }
-
-    private Cookie createCookie(String key, String value) {
-
-        Cookie cookie = new Cookie(key, value);
-        cookie.setMaxAge(24*60*60);
-//        cookie.setSecure(true);
-//        cookie.setPath("/");
-        cookie.setHttpOnly(true);
-
-        return cookie;
+        return bCryptPasswordEncoder.matches(changePwdDTO.getOriginalPwd(), member.getPassword());
     }
 
     @Override
@@ -405,6 +126,22 @@ public class MemberCommandServiceImpl implements MemberCommandService {
         int success = memberRepository.updatePasswordByUserId(userId, bCryptPasswordEncoder.encode(code));
 
         return success != 0;
+    }
+
+    private Boolean checkSmsVerify(String phoneNumber) {
+
+        String savedCode = redisUtils.getSmsAuthCode(phoneNumber);
+        String verifyCode = redisUtils.getSmsAuthCodeVerify(phoneNumber);
+
+        if (savedCode == null || verifyCode == null) {
+            throw new MemberBadRequestException("인증코드를 찾을 수 없습니다.");
+        }
+
+        if (!savedCode.equals(verifyCode)) {
+            throw new MemberBadRequestException("인증코드가 일치하지 않습니다.");
+        }
+
+        return true;
     }
 
     @Override
