+++ src/main/java/com/example/everguide/service/signup/SignupService.java
@@ -1,21 +1,19 @@
-package com.example.everguide.service.member;
+package com.example.everguide.service.signup;
 
-import com.example.everguide.domain.Member;
 import com.example.everguide.web.dto.MemberRequest;
 import com.example.everguide.web.dto.MemberResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.validation.Errors;
 
-import java.util.Map;
-
-public interface MemberQueryService {
-
-    Map<String, String> validateHandling(Errors errors);
+public interface SignupService {
 
     Boolean checkEmailExist(String userId);
 
     MemberResponse.AdditionalNotValidateDTO checkInfoEqual(HttpServletRequest request, HttpServletResponse response, MemberRequest.SignupAdditionalDTO signupAdditionalDTO);
 
-    Boolean checkOriginalPwd(MemberRequest.ChangePwdDTO changePwdDTO);
+    boolean localSignUp(MemberRequest.SignupDTO signupDTO);
+
+    MemberResponse.SignupAdditionalDTO getSignupAdditionalInfo(HttpServletRequest request, HttpServletResponse response);
+
+    boolean registerSignupAdditionalInfo(HttpServletRequest request, HttpServletResponse response, MemberRequest.SignupAdditionalDTO signupAdditionalDTO);
 }
