+++ src/main/java/com/example/everguide/service/token/TokenService.java
@@ -1,0 +1,11 @@
+package com.example.everguide.service.token;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+public interface TokenService {
+
+    Boolean cookieToHeader(HttpServletRequest request, HttpServletResponse response);
+
+    boolean reissue(HttpServletRequest request, HttpServletResponse response);
+}
