+++ src/main/resources/errors.properties
@@ -12,7 +12,7 @@ invalid.email=\uC774\uBA54\uC77C \uD615\uC2DD\uC774 \uC62C\uBC14\uB974\uC9C0 \uC
 
 blank.password=\uBE44\uBC00\uBC88\uD638\uB294 \uD544\uC218 \uC785\uB825 \uAC12\uC785\uB2C8\uB2E4.
 invalid.password=\uBE44\uBC00\uBC88\uD638\uB294 8\uC790 \uC774\uC0C1 \uC601\uBB38, \uC22B\uC790, \uD2B9\uC218\uBB38\uC790\uB97C \uC0AC\uC6A9\uD558\uC138\uC694.
-same.password=\uAE30\uC874 \uBE44\uBC00\uBC88\uD638\uC640 \uC2E0\uADDC \uBE44\uBC00\uBC88\uD638\uAC00 \uAC19\uC2B5\uB2C8\uB2E4.
+duplicate.password=\uAE30\uC874 \uBE44\uBC00\uBC88\uD638\uC640 \uC2E0\uADDC \uBE44\uBC00\uBC88\uD638\uAC00 \uAC19\uC2B5\uB2C8\uB2E4.
 
 blank.rewritePassword=\uBE44\uBC00\uBC88\uD638 \uC7AC\uC785\uB825\uC740 \uD544\uC218 \uC785\uB825 \uAC12\uC785\uB2C8\uB2E4.
 invalid.rewritePassword=\uBE44\uBC00\uBC88\uD638\uC640 \uC7AC\uC785\uB825\uC774 \uAC19\uC9C0 \uC54A\uC2B5\uB2C8\uB2E4.
